move agent into its own module
