+++ agent/tools/ask_human_tool.py
@@ -3,7 +3,7 @@
 import json
 from typing import Dict, Any
 
-from tools.base_tool import ToolDefinition
+from agent.tools.base_tool import ToolDefinition
 
 
 def get_user_message():
